Move Select to concurrency.py; incorporate outputs
