+++ paddle/fluid/operators/select_op.cc
@@ -27,6 +27,7 @@ namespace operators {
 
 static constexpr char kX[] = "X";
 static constexpr char kCaseToExecute[] = "case_to_execute";
+static constexpr char kOutputs[] = "Out";
 
 static constexpr char kCases[] = "cases";
 static constexpr char kCasesBlock[] = "sub_block";
@@ -387,6 +388,10 @@ class SelectOpMaker : public framework::OpProtoAndCheckerMaker {
     AddInput(kCaseToExecute,
              "(Int) The variable the sets the index of the case to execute, "
              "after evaluating the channels being sent to and received from")
+        .AsDuplicable();
+    AddOutput(kOutputs,
+              "A set of variables, which will be assigned with values "
+              "generated by the operators inside the cases of Select Op.")
         .AsDuplicable();
     AddAttr<std::vector<std::string>>(kCases,
                                       "(String vector) Serialized list of"
